Move Asn class into root lib folder
