+++ packages/core/phase1/lib/isAsnValid.ts
@@ -1,6 +1,6 @@
 import type { OrganisationUnitCodes } from "../../types/AnnotatedHearingOutcome"
 
-import Asn from "../lib/Asn"
+import Asn from "../../lib/Asn"
 import isOrganisationUnitValid from "../lib/isOrganisationUnitValid"
 
 const convertAsnToOrganisationUnit = (asn: string): OrganisationUnitCodes => {
